requirements.txt.txt
streamlit
pandas==2.2.2
requests==2.32.3
reportlab==4.2.5  # For PDF generation
python-docx==1.1.2  # For DOCX file processing
pdfplumber==0.11.4